Time: 0 ms (100%), Space: 8.6 MB (18.81%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.cpp
@@ -1,17 +1,16 @@
 class Solution {
 public:
-    int reverse(int x) {
+    long long reverse(int x) {
         long result = 0;
         while (x != 0) {
             int digit = x % 10;
-            result = result * 10 + digit;
+            result = result * 10 + digit; // wow!!
             x /= 10;
 
-            // Check for overflow
             if (result > INT_MAX || result < INT_MIN) {
                 return 0; // Return 0 in case of overflow
             }
         }
-        return result;
+        return (long long)result;
     }
 };
